+++ sample/src/main/java/com/barista_v/sample/MainActivity.java
@@ -4,15 +4,19 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.widget.Toast;
 import com.barista_v.debugging.DebugDrawer;
 import com.barista_v.debugging.ViewServer;
+import com.barista_v.debugging.item.input.InputItemListener;
+import com.barista_v.debugging.item.phoenix.RestartListener;
 import com.barista_v.debugging.item.spinner.SpinnerItemListener;
 import com.jakewharton.scalpel.ScalpelFrameLayout;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class MainActivity extends AppCompatActivity implements SpinnerItemListener {
+public class MainActivity extends AppCompatActivity implements SpinnerItemListener,
+    RestartListener, InputItemListener {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -24,11 +28,13 @@ public class MainActivity extends AppCompatActivity implements SpinnerItemListen
       setSupportActionBar(toolbar);
     }
 
-    String[] hosts = new String[] { "http://asdasdasdasdasdasdasdasd.com", "http://445444.com" };
+    String[] hosts = new String[] { "Value 1", "Value 2" };
     ViewServer.get(this).addWindow(this);
     new DebugDrawer(MyApplication.sInstance, this)
         .withScalpelLayout((ScalpelFrameLayout) findViewById(R.id.scalpel))
-        .withSpinnerItem(1, "Host", hosts, this)
+        .withInputItem(2, "Host", this)
+        .withRestartListener(this)
+        .withSpinnerItem(1, "Spinner Value", hosts, this)
         .withProperties(getProperties())
         .openDrawer();
   }
@@ -58,4 +64,23 @@ public class MainActivity extends AppCompatActivity implements SpinnerItemListen
   public void onItemClick(int itemId, CharSequence title) {
     Toast.makeText(this, "Selected: " + title, Toast.LENGTH_LONG).show();
   }
+
+  //<editor-fold desc="RestartListener">
+  @Override
+  public void onAppRestart() {
+    Log.d("DEBUG", "BOOM App");
+  }
+
+  @Override
+  public void onActivityRestart() {
+    Log.d("DEBUG", "BOOM Activity");
+  }
+  //</editor-fold>
+
+  //<editor-fold desc="InputItemListener">
+  @Override
+  public void onOkClick(int itemId, String inputText) {
+    Toast.makeText(this, inputText, Toast.LENGTH_LONG).show();
+  }
+  //</editor-fold>
 }